+++ src/c/breseq/identify_mutations.cpp
@@ -712,7 +712,7 @@ void identify_mutations_pileup::at_target_end(const uint32_t tid) {
   update_unknown_intervals(target_length(tid)+1, tid, true, false);
 
   // if this target failed to have its coverage fit, mark the entire thing as a deletion
-  uint32_t _this_deletion_propagation_cutoff = _deletion_propagation_cutoffs[tid];
+  uint32_t _this_deletion_propagation_cutoff = static_cast<uint32_t>(_deletion_propagation_cutoffs[tid]);
   
   // if the propagation cutoff is zero then the coverage distribution failed
   if (_this_deletion_propagation_cutoff == 0)
